small move to common
